# horrible quadratic baggage