--- a/typing/jkind.mli
+++ b/typing/jkind.mli
@@ -330,7 +330,7 @@ val add_nullability_crossing : 'd t -> 'd t
 
 (** Take an existing [jkind_l] and add some baggage. *)
 val add_baggage :
-  ?modality:Mode.Modality.Value.Const.t ->
+  modality:Mode.Modality.Value.Const.t ->
   baggage:Types.type_expr ->
   jkind_l ->
   jkind_l