--- a/typing/jkind_axis.mli
+++ b/typing/jkind_axis.mli
@@ -69,6 +69,8 @@ module Axis : sig
 
   val is_deep : _ t -> bool
 
+  (* CR layouts v2.8: Not sure this belongs here, but there's not another obvious spot. Once this
+     file is more aligned with axis treatment in mode.ml, possibly re-home this. *)
   val modality_is_const_for_axis : _ t -> Mode.Modality.Value.Const.t -> bool
 end