@glittershark 2024-12-29
at least insofar as the compiler test suite is a reasonable performance benchmark for the compiler, this is... not actually terrible?

before:
```
user=697.35s system=219.18s cpu=1161% total=1:18.89
```

after:
```
user=812.91s system=217.03s cpu=891% total=1:55.51
```