@stedolan 2025-01-27
> at least insofar as the compiler test suite is a reasonable performance benchmark for the compiler

for future reference: it really, really isn't a good perf benchmark for the compiler. It's dominated by weird things like e.g. regression tests for old obscure GC bugs that compact a million times. the compiler _build_ (main, when using the new compiler) is a better benchmark of compiler performance.